+++ src/lib/components/dropdown/index.ts
@@ -1,1 +1,1 @@
-export { default as DropdownSingle } from './DropdownSingle';
+export { default as DropdownTextfield } from './DropdownTextfield';
